+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Space Invaders
 - Pac-Man
 
+## 2025-03-07
+
+- ADD: Implement Minesweeper game
+
 ## 2025-03-06
 
 - INIT: Initialize project structure with essential configuration files and documentation.